requirements.txt
aioflask==0.4.0
aiosqlite==0.17.0
asgiref==3.5.0
buildozer==1.3.0
click==8.1.2
distlib==0.3.4
filelock==3.6.0
Flask==2.1.1
greenlet==1.1.2
greenletio==0.9.0
h11==0.13.0
importlib-metadata==4.11.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pexpect==4.8.0
platformdirs==2.5.2
ptyprocess==0.7.0
pypika-tortoise==0.1.4
pytz==2022.1
sh==1.14.2
six==1.16.0
tortoise-orm==0.19.0
typing-extensions==4.2.0
uvicorn==0.17.6
virtualenv==20.14.1
Werkzeug==2.1.1
zipp==3.8.0
